deploy/requirements.txt
appnope==0.1.0
boto==2.38.0
boto3==1.3.1
botocore==1.4.40
cycler==0.10.0
decorator==4.0.2
Django==1.9.2
django-environ==0.4.0
docutils==0.12
ecdsa==0.13
Fabric==1.10.2
fabric-slack-tools==0.1
fabtools==0.19.0
futures==3.0.5
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jmespath==0.9.0
lz4==0.8.2
lz4tools==1.3.1.2
MarkupSafe==0.23
matplotlib==1.5.1
meld3==1.0.2
numpy==1.10.4
paramiko==1.15.2
path.py==8.1
pexpect==3.3
pickleshare==0.5
pycrypto==2.6.1
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2015.7
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
supervisor==3.1.3
traitlets==4.0.0
wheel==0.24.0